+++ packages/amplify-provider-awscloudformation/src/__tests__/utils/amplify-resource-state-utils.test.ts
@@ -1,45 +1,49 @@
+import { CloudFormationClient, DescribeStackResourcesCommand, DescribeStacksCommand } from '@aws-sdk/client-cloudformation';
 import { getTableNames, getPreviousDeploymentRecord } from '../../utils/amplify-resource-state-utils';
-import { CloudFormation } from 'aws-sdk';
+import { mockClient } from 'aws-sdk-client-mock';
+import 'aws-sdk-client-mock-jest';
 
-const cfnClientStub = {
-  describeStackResources: () => ({
-    promise: () =>
-      Promise.resolve({
-        StackResources: [
+const mockCfnClient = mockClient(CloudFormationClient);
+mockCfnClient
+  .on(DescribeStackResourcesCommand)
+  .resolves({
+    StackResources: [
+      {
+        LogicalResourceId: 'LogicalResourceIdTest1',
+        PhysicalResourceId: 'PhysicalResourceIdTest',
+        ResourceType: undefined,
+        Timestamp: undefined,
+        ResourceStatus: undefined,
+      },
+    ],
+  })
+  .on(DescribeStacksCommand)
+  .resolves({
+    Stacks: [
+      {
+        Outputs: [
           {
-            LogicalResourceId: 'LogicalResourceIdTest1',
-            PhysicalResourceId: 'PhysicalResourceIdTest',
+            OutputKey: 'GetAttLogicalResourceIdTest1TableName',
+            OutputValue: 'TestStackOutputValue1',
+          },
+          {
+            OutputKey: 'InvalidLogicalResourceIdTableName',
+            OutputValue: 'TestStackOutputValue2',
           },
         ],
-      }),
-  }),
-  describeStacks: () => ({
-    promise: () =>
-      Promise.resolve({
-        Stacks: [
+        Parameters: [
           {
-            Outputs: [
-              {
-                OutputKey: 'GetAttLogicalResourceIdTest1TableName',
-                OutputValue: 'TestStackOutputValue1',
-              },
-              {
-                OutputKey: 'InvalidLogicalResourceIdTableName',
-                OutputValue: 'TestStackOutputValue2',
-              },
-            ],
-            Parameters: [
-              {
-                ParameterKey: 'TestParameterKey1',
-                ParameterValue: 'TestParameterValue1',
-              },
-            ],
-            Capabilities: ['CAPABILITY_IAM'],
+            ParameterKey: 'TestParameterKey1',
+            ParameterValue: 'TestParameterValue1',
           },
         ],
-      }),
-  }),
-} as unknown as CloudFormation;
+        Capabilities: ['CAPABILITY_IAM'],
+        StackName: undefined,
+        CreationTime: undefined,
+        StackStatus: undefined,
+      },
+    ],
+  });
 
 describe('amplify-resource-state-utils', () => {
   const StackID = 'TestSTackID';
@@ -49,7 +53,7 @@ describe('amplify-resource-state-utils', () => {
     const expectedTableNameMap: Map<string, string> = new Map();
     expectedTableNameMap.set('LogicalResourceIdTest1', 'TestStackOutputValue1');
 
-    const tableNames = await getTableNames(cfnClientStub, tables, StackID);
+    const tableNames = await getTableNames(mockCfnClient as unknown as CloudFormationClient, tables, StackID);
     expect(tableNames).toEqual(expectedTableNameMap);
   });
 
@@ -58,7 +62,7 @@ describe('amplify-resource-state-utils', () => {
       capabilities: ['CAPABILITY_IAM'],
       parameters: { TestParameterKey1: 'TestParameterValue1' },
     };
-    const prevDeploymentRecord = await getPreviousDeploymentRecord(cfnClientStub, StackID);
+    const prevDeploymentRecord = await getPreviousDeploymentRecord(mockCfnClient as unknown as CloudFormationClient, StackID);
 
     expect(prevDeploymentRecord).toEqual(expectedPrevDeploymentRecord);
   });