+++ packages/amplify-provider-awscloudformation/src/aws-utils/aws-location-service.ts
@@ -1,16 +1,16 @@
 import { $TSContext } from '@aws-amplify/amplify-cli-core';
-import * as AWS from 'aws-sdk';
-import { AwsSecrets, loadConfiguration } from '../configuration-manager';
-import aws from './aws.js';
+import { LocationClient } from '@aws-sdk/client-location';
+import { NodeHttpHandler } from '@smithy/node-http-handler';
+import { loadConfiguration } from '../configuration-manager';
 import { proxyAgent } from './aws-globals';
 
 export class LocationService {
   private static instance: LocationService;
-  readonly client: AWS.Location;
+  readonly client: LocationClient;
 
   static async getInstance(context: $TSContext, options = {}): Promise<LocationService> {
     if (!LocationService.instance) {
-      let cred: AwsSecrets = {};
+      let cred = {};
       try {
         cred = await loadConfiguration(context);
       } catch (e) {
@@ -21,13 +21,14 @@ export class LocationService {
     return LocationService.instance;
   }
 
-  private constructor(cred: AwsSecrets, options = {}) {
-    this.client = new aws.Location({
+  private constructor(cred, options = {}) {
+    this.client = new LocationClient({
       ...cred,
       ...options,
-      httpOptions: {
-        agent: proxyAgent(),
-      },
+      requestHandler: new NodeHttpHandler({
+        httpAgent: proxyAgent(),
+        httpsAgent: proxyAgent(),
+      }),
     });
   }
 }