+++ packages/amplify-util-mock/src/utils/lambda/populate-lambda-mock-env-vars.ts
@@ -31,9 +31,9 @@ const getAwsCredentials = async (__, context: $TSContext): Promise<Record<string
   }
   const awsConfigInfo = await loadConfigurationForEnv(context, env, appId);
   return {
-    AWS_ACCESS_KEY_ID: awsConfigInfo.accessKeyId,
-    AWS_SECRET_ACCESS_KEY: awsConfigInfo.secretAccessKey,
-    AWS_SESSION_TOKEN: awsConfigInfo.sessionToken,
+    AWS_ACCESS_KEY_ID: awsConfigInfo.credentials.accessKeyId,
+    AWS_SECRET_ACCESS_KEY: awsConfigInfo.credentials.secretAccessKey,
+    AWS_SESSION_TOKEN: awsConfigInfo.credentials.sessionToken,
   };
 };