+++ packages/amplify-nodejs-function-template-provider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.10.16](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-nodejs-function-template-provider@2.10.15...@aws-amplify/amplify-nodejs-function-template-provider@2.10.16) (2025-10-08)
+
+**Note:** Version bump only for package @aws-amplify/amplify-nodejs-function-template-provider
+
+
+
+
+
 ## [2.10.15](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-nodejs-function-template-provider@2.10.14...@aws-amplify/amplify-nodejs-function-template-provider@2.10.15) (2025-07-01)
 
 **Note:** Version bump only for package @aws-amplify/amplify-nodejs-function-template-provider