+++ packages/amplify-provider-awscloudformation/CHANGELOG.md
@@ -3,6 +3,56 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.11.13](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-provider-awscloudformation@8.11.12...@aws-amplify/amplify-provider-awscloudformation@8.11.13) (2025-10-16)
+
+
+### Bug Fixes
+
+* update throttling retry ([#14280](https://github.com/aws-amplify/amplify-cli/issues/14280)) ([f461d5f](https://github.com/aws-amplify/amplify-cli/commit/f461d5fc669822b835e23e2df3b260bafaa5e3bb))
+
+
+
+
+
+## [8.11.12](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-provider-awscloudformation@8.11.11...@aws-amplify/amplify-provider-awscloudformation@8.11.12) (2025-10-10)
+
+
+### Bug Fixes
+
+* `amplify codegen` command ([#14277](https://github.com/aws-amplify/amplify-cli/issues/14277)) ([53e973b](https://github.com/aws-amplify/amplify-cli/commit/53e973b0659ebad2728b82ed38e0adea6071aa59))
+
+
+
+
+
+## [8.11.11](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-provider-awscloudformation@8.11.10...@aws-amplify/amplify-provider-awscloudformation@8.11.11) (2025-10-08)
+
+**Note:** Version bump only for package @aws-amplify/amplify-provider-awscloudformation
+
+
+
+
+
+## [8.11.10](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-provider-awscloudformation@8.11.9...@aws-amplify/amplify-provider-awscloudformation@8.11.10) (2025-10-01)
+
+
+### Bug Fixes
+
+* e2e tests for auth, geo, and python lambda functions ([#14264](https://github.com/aws-amplify/amplify-cli/issues/14264)) ([da6de92](https://github.com/aws-amplify/amplify-cli/commit/da6de92bf764db9f9c2797fbd1fd1d0f5b9c578a))
+* Userpool does not exist ([#14262](https://github.com/aws-amplify/amplify-cli/issues/14262)) ([5464abe](https://github.com/aws-amplify/amplify-cli/commit/5464abe68d55faf0b91d4306f119a844abdc6d3d))
+
+
+
+
+
+## [8.11.9](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-provider-awscloudformation@8.11.8...@aws-amplify/amplify-provider-awscloudformation@8.11.9) (2025-09-17)
+
+**Note:** Version bump only for package @aws-amplify/amplify-provider-awscloudformation
+
+
+
+
+
 ## [8.11.8](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-provider-awscloudformation@8.11.7...@aws-amplify/amplify-provider-awscloudformation@8.11.8) (2025-07-01)